requirements.txt
fastapi~=0.115.14
langchain~=0.3.26
python-dotenv~=1.1.1
requests~=2.32.4
streamlit~=1.46.1